Добавляет setTimeout в FilmCard
