+++ project/src/components/film-list/film-list.jsx
@@ -11,7 +11,7 @@ function FilmList({films}) {
 
   return (
     <div className="catalog__films-list">
-      {films.map((item) => <FilmCard film={item}  key={item.id} isPlaying={(activeFilm.id === item.id)}  onCardHover={changeActiveFilmHandler}/>)}
+      {films.map((item) => <FilmCard film={item}  key={item.id} isPlaying={(activeFilm.id === item.id)} onCardHover={changeActiveFilmHandler}/>)}
     </div>
   );
 }
@@ -21,3 +21,25 @@ FilmList.propTypes = {
 };
 
 export default FilmList;
+
+
+/*
+useEffect(() => {
+    const hoverTimer = setTimeout(() => {
+      if (isPlaying && videoRef.current) {
+        videoRef.current.play();
+      }
+    }, VIDEO_DELAY);
+
+    return () => {
+      if (videoRef.current) {
+        videoRef.current.pause();
+        //videoRef.current.currentTime = 0;
+
+        videoRef.current.src=filmVideo;
+      }
+      clearTimeout(hoverTimer);
+
+    }
+  }, [isPlaying]);
+ */
